Renamed from plural to single form
