+++ CargoHubV2/app/models/item_types_model.py
@@ -14,4 +14,3 @@ class ItemType(Base):
     updated_at = Column(DateTime, default=datetime.now, onupdate=datetime.now)
 
     items = relationship("Item", back_populates="item_type_rel")
-
