+++ CargoHubV2/app/schemas/item_groups_schema.py
